add event handling
After much trying with generics and question marks and catches I came to the conclusion that it wouldn't be possible to do in a reasonable fashion. As a bonus a command handling now allows handling of multiple commands in the same function. Although that is not recommended usage unless multiple commands have very similar results.